requirements.txt
fastapi==0.97.0
python-multipart==0.0.6
pypdf==3.10.0
fpdf2==2.7.4
pycryptodome==3.18.0
uvicorn
